requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
pillow==12.2.0
    # via -r requirements.in
